# backend/api/openapi/api.yaml
openapi: 3.0.3
info:
  title: School API
  description: API для платформы онлайн/оффлайн школы и помощника репетитора.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Основной шлюз API
paths:
...
  /classrooms:
    get:
      summary: Получение списка учебных кабинетов
      operationId: listClassrooms
      tags:
        - Classrooms
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список кабинетов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassroomResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Создание нового учебного кабинета
      operationId: createClassroom
      tags:
        - Classrooms
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassroomRequest'
      responses:
        '201':
          description: Кабинет успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassroomResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /lessons:
    get:
      summary: Список уроков с фильтрацией
      operationId: listLessons
      tags:
        - Lessons
      security:
        - BearerAuth: []
      parameters:
        - name: teacher_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: client_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: classroom_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/LessonStatus'
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Список уроков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LessonResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Назначение урока преподавателем
      operationId: createLesson
      tags:
        - Lessons
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLessonRequest'
      responses:
        '201':
          description: Урок успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Коллизия кабинета или наложение времени
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /lessons/{id}/complete:
    post:
      summary: Завершение проведенного урока преподавателем
      operationId: completeLesson
      tags:
        - Lessons
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Урок успешно завершен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonResponse'
        '400':
          description: Некорректный статус урока для завершения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Урок не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /lessons/{id}/cancel:
    post:
      summary: Отмена урока
      operationId: cancelLesson
      tags:
        - Lessons
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelLessonRequest'
      responses:
        '200':
          description: Урок успешно отменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonResponse'
        '400':
          description: Некорректный статус урока для отмены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Урок не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Введите access_token в формате Bearer <token>
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input parameters
        details:
          type: array
          items:
            type: object
            additionalProperties: true
    ClassroomResponse:
      type: object
      required:
        - id
        - name
        - capacity
        - color
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        name:
          type: string
          example: 'Кабинет №1 (Физика)'
        capacity:
          type: integer
          example: 4
        color:
          type: string
          example: '#3B82F6'
        description:
          type: string
          nullable: true
          example: 'Оснащен маркерной доской и микроскопом'
        created_at:
          type: string
          format: date-time
          example: '2026-10-04T12:00:00Z'
    CreateClassroomRequest:
      type: object
      required:
        - name
        - capacity
      properties:
        name:
          type: string
          minLength: 2
          example: 'Кабинет №1'
        capacity:
          type: integer
          minimum: 1
          example: 4
        color:
          type: string
          nullable: true
          example: '#3B82F6'
        description:
          type: string
          nullable: true
          example: 'Основной кабинет математики'
    LessonFormat:
      type: string
      enum:
        - online
        - offline
      example: offline
    LessonStatus:
      type: string
      enum:
        - scheduled
        - completed
        - cancelled
      example: scheduled
    LessonResponse:
      type: object
      required:
        - id
        - teacher_id
        - client_id
        - start_time
        - end_time
        - format
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        teacher_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174001'
        client_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174002'
        classroom_id:
          type: string
          format: uuid
          nullable: true
          example: '123e4567-e89b-12d3-a456-426614174003'
        start_time:
          type: string
          format: date-time
          example: '2026-10-05T10:00:00Z'
        end_time:
          type: string
          format: date-time
          example: '2026-10-05T11:00:00Z'
        format:
          $ref: '#/components/schemas/LessonFormat'
        location_or_url:
          type: string
          nullable: true
          example: 'Кабинет №1'
        status:
          $ref: '#/components/schemas/LessonStatus'
        notes:
          type: string
          nullable: true
          example: 'Подготовить ДЗ №3'
        cancel_reason:
          type: string
          nullable: true
          example: 'Ученик заболел'
        created_at:
          type: string
          format: date-time
          example: '2026-10-04T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-10-04T12:00:00Z'
    CreateLessonRequest:
      type: object
      required:
        - client_id
        - start_time
        - end_time
        - format
      properties:
        client_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174002'
        classroom_id:
          type: string
          format: uuid
          nullable: true
          example: '123e4567-e89b-12d3-a456-426614174003'
        start_time:
          type: string
          format: date-time
          example: '2026-10-05T10:00:00Z'
        end_time:
          type: string
          format: date-time
          example: '2026-10-05T11:00:00Z'
        format:
          $ref: '#/components/schemas/LessonFormat'
        location_or_url:
          type: string
          nullable: true
          example: 'https://zoom.us/j/123456789'
        notes:
          type: string
          nullable: true
          example: 'Тема: Квадратные уравнения'
    CancelLessonRequest:
      type: object
      properties:
        reason:
          type: string
          nullable: true
          example: 'Форс-мажор'
...